+++ src/main/java/org/openshapa/views/continuous/sound/SoundDataViewer.java
@@ -152,7 +152,8 @@ public final class SoundDataViewer extends JFrame
     public long getDuration() {
         try {
             if (audio != null) {
-                return Constants.TICKS_PER_SECOND * audio.getDuration() / audio.getTimeScale();
+                return (long)Constants.TICKS_PER_SECOND *
+                        (long)audio.getDuration() / (long)audio.getTimeScale();
             }
         } catch (StdQTException ex) {
             logger.error("Unable to determine QT audio duration", ex);
